requirements.txt
fastapi==0.118.0
google-genai==1.39.1
